+++ microns_utils/filepath_utils.py
@@ -5,7 +5,7 @@ import os
 from pathlib import Path
 import datetime
 import logging
-import pytz
+from .datetime_utils import timezone_converter
 
 def find_all_matching_files(name, path):
     """
@@ -26,26 +26,6 @@ def validate_filepath(filepath):
     filepath = Path(filepath)
     assert filepath.exists()
     return filepath
-
-
-def timezone_converter(timestamp, source_tz, destination_tz, fmt=None):
-    """
-    Converts timestamp from a source timezone to a destination timezone. 
-    To see commonly used timezone options run: 
-    
-    ```from pytz import common_timezones```
-    
-    :param timestamp: (datetime.datetime) timestamp to convert
-    :param source_tz: (pytz.timezone) source timezone to convert 
-    :param destination_tz: (pytz.timezone) destination timezone
-    :param fmt: (str) optional - timestamp format to pass to strftime
-    
-    :returns: (datetime.datetime) converted timestamp
-    """
-    dest_obj = pytz.timezone(destination_tz)
-    source_obj = pytz.timezone(source_tz).localize(timestamp.replace(tzinfo=None))
-    converted = dest_obj.normalize(source_obj.astimezone(dest_obj))
-    return converted if fmt is None else converted.strftime(fmt)
 
 
 def get_file_modification_time(filepath, timezone, fmt=None):
